+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Change Log
 
+### 0.0.3
+* Move json recipes to java code fixes GTCE-Inventory/GTCE-Inventory#1 (warjort)
+* Remove old (and no longer supported by GTCE) forge multipart (warjort)
+* Update to GTCE 1.17.1.770 (warjort)
+
 ### 0.0.2
 * Fix a bug where energy was charged twice because it was being charged in simulate mode (warjort)